requirements.txt
autopep8==2.0.1
black==23.1.0
bluesky-guidata==2022.9.15
bluesky-simdata==2022.9.15
bluesky-simulator==2022.12.22
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
filelock==3.9.0
fonttools==4.38.0
geographiclib==2.0
geopy==2.3.0
importlib-metadata==4.13.0
kiwisolver==1.4.4
llvmlite==0.39.1
markdown-it-py==2.1.0
matplotlib==3.7.0
mdurl==0.1.2
msgpack==1.0.4
mypy-extensions==1.0.0
nanoid==2.0.0
numba==0.56.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
pipenv==2023.2.4
platformdirs==2.6.2
pycodestyle==2.10.0
pygame==2.1.3
Pygments==2.14.0
pymap3d==2.9.1
PyOpenGL==3.1.6
pyparsing==3.0.9
pyproj==3.4.1
PyQt6==6.4.2
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
PyQt6-WebEngine==6.4.0
PyQt6-WebEngine-Qt6==6.4.2
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
rich==13.3.1
scipy==1.10.0
shapely==2.0.1
simplekml==1.3.6
six==1.16.0
textual==0.10.1
tomli==2.0.1
tqdm==4.64.1
virtualenv==20.19.0
virtualenv-clone==0.5.7
zipp==3.13.0
zmq==0.0.0
